+++ SimpleSDMDatasets/src/SimpleSDMDatasets.jl
@@ -4,6 +4,7 @@ import Downloads
 using Dates
 import ZipArchives
 using Markdown
+using TestItems
 
 # Set the potential paths for downloads
 const _data_storage_folders = first([